feat : 프론트 형성평가 미리보기 추가
형성평가 미리보기 모달 기능 추가

+++ Backend/edututor/src/main/java/com/tls/edututor/exam/sharetest/repository/ShareTestRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ShareTestRepository extends JpaRepository<ShareTest, Integer> {
+public interface ShareTestRepository extends JpaRepository<ShareTest, Long> {
 
   Optional<ShareTest> findByUserIdAndTestPaperId(Long userId, Long testPaperId);
